Refactor if/else structure to foreach
